backend/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy>=1.4 # Dobré špecifikovať minimálnu verziu
Flask-CORS
Flask-Marshmallow
marshmallow-sqlalchemy
python-dotenv